refactor: 의미가 모호한 서비스 인터페이스 제거
