+++ src/main/java/roomescape/reservation/application/service/ReservationViewQueryService.java
@@ -1,16 +1,31 @@
 package roomescape.reservation.application.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import roomescape.reservation.application.dto.CreateReservationServiceRequest;
 import roomescape.reservation.domain.ReservationView;
+import roomescape.reservation.domain.ReservationViewRepository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface ReservationViewQueryService {
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ReservationViewQueryService {
 
-    boolean existsByParams(CreateReservationServiceRequest serviceRequest, Long id);
+    private final ReservationViewRepository viewRepository;
 
-    List<ReservationView> getAllByUserId(Long userId);
+    public List<ReservationView> getAllByUserId(final Long userId) {
+        return viewRepository.findAllByUserId(userId);
+    }
 
-    Optional<Long> findFirstWaitingByReservationId(Long id);
+    public boolean existsByParams(final CreateReservationServiceRequest request, final Long userId) {
+        return viewRepository.existsByParams(request.date(), request.timeId(), request.themeId(), userId);
+    }
+
+    public Optional<Long> findFirstWaitingByReservationId(final Long id) {
+        return viewRepository.findFirstWaitingByReservationId(id);
+    }
 }
