+++ src/main/java/roomescape/theme/application/service/ThemeQueryService.java
@@ -1,20 +1,48 @@
 package roomescape.theme.application.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.common.domain.DomainTerm;
+import roomescape.common.exception.NotFoundException;
+import roomescape.reservation.application.dto.ThemeToBookCountServiceResponse;
+import roomescape.reservation.application.service.ReservationQueryService;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeName;
+import roomescape.theme.domain.ThemeRepository;
 
 import java.util.List;
 
-public interface ThemeQueryService {
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ThemeQueryService {
 
-    boolean existsById(Long id);
+    private final ThemeRepository themeRepository;
+    private final ReservationQueryService reservationQueryService;
 
-    boolean existsByName(ThemeName name);
+    public boolean existsById(final Long id) {
+        return themeRepository.existsById(id);
+    }
 
-    List<Theme> getAll();
+    public boolean existsByName(final ThemeName name) {
+        return themeRepository.existsByName(name);
+    }
 
-    Theme get(Long id);
+    public List<Theme> getAll() {
+        return themeRepository.findAll();
+    }
 
-    List<Theme> getRanking(ReservationDate startDate, ReservationDate endDate, int count);
+    public Theme get(final Long id) {
+        return themeRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException(DomainTerm.THEME, id));
+    }
+
+    public List<Theme> getRanking(final ReservationDate startDate, final ReservationDate endDate, final int count) {
+        return reservationQueryService.getRanking(startDate, endDate, count).stream()
+                .map(ThemeToBookCountServiceResponse::theme)
+                .toList();
+    }
+
 }
