+++ src/test/java/roomescape/time/application/service/ReservationTimeQueryServiceTest.java
@@ -15,10 +15,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Transactional
-class ReservationTimeQueryServiceImplTest {
+class ReservationTimeQueryServiceTest {
 
     @Autowired
-    private ReservationTimeQueryServiceImpl reservationTimeQueryService;
+    private ReservationTimeQueryService reservationTimeQueryService;
 
     @Autowired
     private ReservationTimeRepository reservationTimeRepository;
